+++ test/web/twitter_api/views/user_view_test.exs
@@ -112,9 +112,11 @@ defmodule Pleroma.Web.TwitterAPI.UserViewTest do
     as_user = UserView.render("show.json", %{user: user, for: user})
     assert as_user["default_scope"] == user.info.default_scope
     assert as_user["no_rich_text"] == user.info.no_rich_text
+    assert as_user["pleroma"]["notification_settings"] == user.info.notification_settings
     as_stranger = UserView.render("show.json", %{user: user})
     refute as_stranger["default_scope"]
     refute as_stranger["no_rich_text"]
+    refute as_stranger["pleroma"]["notification_settings"]
   end
 
   test "A user for a given other follower", %{user: user} do